<fix>: createdAt no longer borks
[Why]: When user fetch items with createdAt field, internally we turn it
from ISO 8601 format to Unix time, but if user send it back in verbatim
on an PUT or PUT all without removing the unix time it will bork because
system doesn't recognizes the unix time conversion. And I don't know how
to make it automatic since it's nested inside the code. Also we don't want
the user to update it anyway.

[How]: While taking user input, check the model's cherry-pick permissions
when we do out output using the Admin role. If "createdAt" field exists,
use JSON patch to remove the field in the user input.