<feat>: Models can control endpoint access by defining GuardAPIEntry or override before/after hookpoints
[Why]: Perhaps some token cannot access certain endpoint, so scope,
endpoint, method are given to GuardAPIEntry() which return true or false,
if given. Otherwise by default it's a pass.

If there is a use case where maybe the access depend on the data
themselves, not determined by scope, endpoint, and method triplet,
then model can return error in the before/after hookpoints. Scope is also
given now to these method to allow the model to tailor to the
information.